requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
# Dependencias para autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
